make ActivationMethod an enum class

+++ src/openms/source/FORMAT/FLASHDeconvSpectrumFile.cpp
@@ -329,7 +329,7 @@ namespace OpenMS
       double precursor_mass = dspec.getPrecursorPeakGroup().getMonoMass();
       if (dspec.getActivationMethod() < Precursor::ActivationMethod::SIZE_OF_ACTIVATIONMETHOD)
       {
-        fs << "ACTIVATION=" << Precursor::NamesOfActivationMethodShort[dspec.getActivationMethod()] << "\n";
+        fs << "ACTIVATION=" << Precursor::NamesOfActivationMethodShort[static_cast<int>(dspec.getActivationMethod())] << "\n";
       }
       fs << "MS_ONE_ID=" << dspec.getPrecursorScanNumber() << "\n"
          << "MS_ONE_SCAN=" << dspec.getPrecursorScanNumber() << "\n"
